requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
cloudinary==1.41.0
Django==5.0.7
django-axes==6.5.1
django-cloudinary-storage==0.3.0
django-formtools==2.5.1
django-localflavor==4.0
idna==3.7
pillow==10.4.0
python-dotenv==1.0.1
python-stdnum==1.20
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
